+++ native/zohodesk_portal_configuration/lib/zohodesk_portal_configuration.dart
@@ -1,4 +1,4 @@
-
+import 'package:zohodesk_portal_configuration/data/zdp_configuration.dart';
 import 'zohodesk_portal_configuration_platform_interface.dart';
 import 'package:zohodesk_portal_configuration/data/configuration_model.dart';
 
@@ -7,4 +7,6 @@ class ZohodeskPortalConfiguration {
   static setTheme(ZDPTheme theme) async => await ZohodeskPortalConfigurationPlatform.instance.setTheme(theme);
 
   static handleNotification(Map<String, dynamic> map) async => await ZohodeskPortalConfigurationPlatform.instance.handleNotification(map);
+
+  static setConfiguration(ZDPConfiguration configuration) async => await ZohodeskPortalConfigurationPlatform.instance.setConfiguration(configuration);
 }