+++ native/zohodesk_portal_ticket/android/src/main/kotlin/com/zoho/desk/asap/ticket/zohodesk_portal_ticket/data/ZDPTicketConfiguration.kt
@@ -1,0 +1,14 @@
+package com.zoho.desk.asap.ticket.zohodesk_portal_ticket.data
+
+data class ZDPTicketConfiguration (
+    val isReplyAllowed: Boolean,
+    val isCommentAllowed: Boolean,
+    val isTicketUpdateAllowed: Boolean,
+    val isCommentEditAllowed: Boolean,
+    val isCommentDeleteAllowed: Boolean,
+    val isAddTicketAllowed: Boolean,
+    val isHappinessThreadAllowed: Boolean,
+    val isTicketPropertiesAllowed: Boolean,
+    val isTicketChannelAllowed: Boolean,
+    val isTicketDetailSearchAllowed: Boolean,
+)