+++ native/zohodesk_portal_ticket/lib/zohodesk_portal_ticket_method_channel.dart
@@ -1,6 +1,7 @@
+import 'dart:convert';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
-
+import 'data/zdp_ticket_configuration.dart';
 import 'zohodesk_portal_ticket_platform_interface.dart';
 
 /// An implementation of [ZohodeskPortalTicketPlatform] that uses method channels.
@@ -12,4 +13,8 @@ class MethodChannelZohodeskPortalTicket extends ZohodeskPortalTicketPlatform {
   ///ASAP Ticket show channel method
   @override
   Future<void> show() async => await methodChannel.invokeMethod<String>('show');
+
+  ///ASAP Ticket set configuration channel method
+  @override
+  Future<void> setConfiguration(ZDPTicketConfiguration configuration) async => await methodChannel.invokeMethod('setConfiguration', {'configuration': jsonEncode(configuration)});
 }