+++ native/zohodesk_portal_ticket/CHANGELOG.md
@@ -59,3 +59,8 @@ ASAP SDK Portal Tickets flutter plugin
 - Integrated Latest version of Android SDK v3.0_BETA_22
 - Added proguard rules for release mode
 - Added namespace in build.gradle for each modules and removed the package name from manifest
+  
+# 1.1.4
+- Integrated the lastest version of iOS ASAP SDK v3.0.12
+- Provided a method to customize the Configuration module
+- Provided a method to customize the Tickets module